[#2] refactor: core 패키지 삭제
